# New and changed contributor scanning improvements