--- a/Slim/Schema/Contributor.pm
+++ b/Slim/Schema/Contributor.pm
@@ -255,14 +255,6 @@ sub rescan {
 		SELECT COUNT(*) FROM contributor_track WHERE contributor = ?
 	} );
 
-	# There may have been more than one Artist tag in the music file, so track.primary_artist does not tell the whole story. So also check for ARTIST/ALBUMARTIST in contributor_track.
-	my $albumSth = $dbh->prepare_cached( qq{
-		SELECT COUNT(1) FROM tracks
-		LEFT JOIN contributor_track ON tracks.id=contributor_track.track
-		WHERE album = ?
-		AND ( tracks.primary_artist = ? OR ( contributor_track.contributor=? AND contributor_track.role IN (1,5) ) )
-	} );
-
 	for my $id ( @$ids ) {
 		$contributorSth->execute($id);
 		my ($count) = $contributorSth->fetchrow_array;
@@ -273,16 +265,6 @@ sub rescan {
 
 			# This will cascade within the database to contributor_album and contributor_track
 			$dbh->do( "DELETE FROM contributors WHERE id = ?", undef, $id );
-		}
-		# contributor->album relations aren't removed automatically when the last track with this primary_artist disappears
-		elsif ( $albumId ) {
-			$albumSth->execute($albumId, $id, $id);
-			($count) = $albumSth->fetchrow_array;
-			$albumSth->finish;
-
-			if ( !$count ) {
-				$dbh->do( "DELETE FROM contributor_album WHERE role IN (1, 5) AND album = ? AND contributor = ?", undef, $albumId, $id );
-			}
 		}
 	}
 }