--- a/Slim/Utils/Scanner/Local.pm
+++ b/Slim/Utils/Scanner/Local.pm
@@ -283,6 +283,16 @@ sub rescan {
 				WHERE  scanned_files.url LIKE '$basedir%'
 		} );
 
+		# and a table of albums being changed
+		$log->error("Build temporary table for changed albums") unless main::SCANNER && $main::progress;
+		$dbh->do('DROP TABLE IF EXISTS changed_albums');
+		$dbh->do( qq{
+			CREATE $createTemporary TABLE changed_albums AS
+				SELECT DISTINCT(tracks.album) AS album
+				FROM   changed
+				JOIN   tracks ON changed.url = tracks.url
+		} );
+
 		my $changedOnlySQL = qq{
 			SELECT url
 			FROM   changed
@@ -314,6 +324,30 @@ sub rescan {
 		}, $args);
 
 		$class->updateTracks($dbh, \$changes, $paths, $next, $changedOnlyCount, $changedOnlySQL, $args);
+
+		# Completely rebuild contributor_album for all changed albums...
+		# ... first, now that tracks has been updated, add albums referenced now which weren't referenced before (in case the user has moved a track to a different album)
+		$log->error("Adding to temporary table for changed albums") unless main::SCANNER && $main::progress;
+		$dbh->do( qq{
+			INSERT INTO changed_albums
+			SELECT DISTINCT(tracks.album) AS album
+			FROM   changed
+			JOIN   tracks ON changed.url = tracks.url
+			WHERE NOT EXISTS (SELECT * FROM changed_albums WHERE changed_albums.album = tracks.album)
+		} );
+		# ... now, rebuild contributor_album
+		$dbh->do( qq{
+			DELETE FROM contributor_album
+			WHERE contributor_album.album IN (SELECT album FROM changed_albums)
+		} );
+		$dbh->do( qq{
+			INSERT INTO contributor_album (role,contributor,album) 
+			SELECT DISTINCT role,contributor, tracks.album 
+			FROM contributor_track
+			JOIN tracks ON tracks.id=contributor_track.track
+			JOIN changed_albums ON tracks.album = changed_albums.album
+		} );
+		main::SCANNER && Slim::Schema->forceCommit;
 
 		$class->addTracks($dbh, \$changes, $paths, $next, $onDiskOnlyCount, $onDiskOnlySQL, $args);