@michaelherger 2024-01-18
Did you notice any performance hit? I fear re-running those full queries might be quite slow compared to some individual updates?