@darrell-k 2024-01-18
We've got an index on tracks(url) so the temporary table creation should be fine. The join to tracks on url is already being done for the existing "changed" table, presumably that is one reason why the index exists. 

I could add an album column to the existing "changed" table rather than creating a separate "changed_albums" table, but that would make  the "SELECT DISTINCT" in  the insertion work a bit harder which might be noticable with a large set of changed tracks, so I don't know that it would be worth it.

Anyway, as I said above, metrics incoming.