@darrell-k 2024-01-18
The results are in. 

Library size: 6508 albums, 86853 tracks

Old:
Media folder scan total (from the log message): 13 changed tracks: 8.1 seconds, 13000 changed tracks: 156 seconds
changed file scan only (diff in log timestamps): 13 changed tracks: 0.4 seconds, 13000 changed tracks: 149 seconds

New:
Media folder scan total (from the log message): 13 changed tracks: 8.2 seconds, 13000 changed tracks: 145 seconds
changed file scan only (diff in log timestamps): 13 changed tracks: 0.5 seconds, 13000 changed tracks: 137 seconds

My conclusion is that there is no significant difference in performance. The test was somewhat biased towards the old method because there were no actual tag changes in the files, meaning that the old method would not have made any changes to the contributor_album table,  whereas the new method rebuilds it regardless.

This is on a (very) cheap Pentium mini PC with 8G RAM running Debian with the music library on a USB-attached mechanical disk, which may account for my relatively prolonged file/folder discovery times - basically all of the 8 seconds shown above!