@mherger 2024-01-25
Let's give that a try! Thanks a lot! I'll tweak the SQL on my end.