@darrell-k 2024-01-25
INSERT OR IGNORE is a new one on me, I was brought up on DB2 SQL which has it's own set of extensions to ANSI SQL but not  that one. 

If you need the unique index to make that work, I guess what "OR IGNORE" is doing is silently surpressing primary key violation errors during the INSERT. I used to write a error handler for that sort of thing!

Aren't standards a wonderful thing?

Anyway, Thanks, hopefully this will cut down a little bit on "my n&c scan didn't work" questions.